Struct cluster_test::cluster::Cluster

pub struct Cluster {
    pub chain_id: ChainId,
    // some fields omitted
}

Fields
chain_id: ChainId
Implementations

impl Cluster

pub fn from_host_port(
peers: Vec<(String, u32, Option<u32>)>,
mint_file: &str,
chain_id: ChainId,
vasp: bool
) -> Self

pub fn new(
validator_instances: Vec<Instance>,
fullnode_instances: Vec<Instance>,
lsr_instances: Vec<Instance>,
vault_instances: Vec<Instance>,
waypoint: Option<Waypoint>
) -> Self

pub fn random_validator_instance(&self) -> Instance

pub fn validator_instances(&self) -> &[Instance]

pub fn random_fullnode_instance(&self) -> Instance

pub fn fullnode_instances(&self) -> &[Instance]

pub fn lsr_instances(&self) -> &[Instance]

pub fn vault_instances(&self) -> &[Instance]

pub fn all_instances(&self) -> impl Iterator<Item = &Instance>

pub fn validator_and_fullnode_instances(
&self
) -> impl Iterator<Item = &Instance>

pub fn into_validator_instances(self) -> Vec<Instance>

pub fn into_fullnode_instances(self) -> Vec<Instance>

pub fn into_lsr_instances(self) -> Vec<Instance>

pub fn into_vault_instances(self) -> Vec<Instance>

pub fn mint_key_pair(&self) -> &KeyPair<Ed25519PrivateKey, Ed25519PublicKey>

pub fn get_validator_instance(&self, name: &str) -> Option<&Instance>

pub fn split_n_validators_random(&self, c: usize) -> (Self, Self)

Splits this cluster into two

Returns tuple of two clusters: First element in tuple contains cluster with c random instances from self Second element in tuple contains cluster with remaining instances from self

pub fn split_n_fullnodes_random(&self, c: usize) -> (Self, Self)

pub fn validator_sub_cluster(&self, ids: Vec<String>) -> Cluster

pub fn find_instance_by_pod(&self, pod: &str) -> Option<&Instance>

pub fn instances_for_group(
&self,
validator_group: ValidatorGroup
) -> impl Iterator<Item = &Instance>

pub fn lsr_instances_for_validators(
&self,
validators: &[Instance]
) -> Vec<Instance>

pub fn vault_instances_for_validators(
&self,
validators: &[Instance]
) -> Vec<Instance>