Trait cluster_test::cluster_swarm::ClusterSwarm

pub trait ClusterSwarm {
#[must_use]
    fn remove_all_network_effects<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn run<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        instance: &'life1 Instance, 
        docker_image: &'life2 str, 
        command: String, 
        job_name: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn validator_instances<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Vec<Instance>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn fullnode_instances<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Vec<Instance>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn upsert_node<'life0, 'async_trait>(
        &'life0 self, 
        instance_config: InstanceConfig, 
        delete_data: bool
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn delete_node<'life0, 'async_trait>(
        &'life0 self, 
        instance_config: InstanceConfig
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn delete_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_grafana_baseurl<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn create_validator_set<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        num_validators: u32, 
        num_fullnodes_per_validator: u32, 
        image_tag: &'life1 str, 
        config_overrides: Vec<String>, 
        delete_data: bool
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

#[must_use]
    fn create_fullnode_set<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        num_validators: u32, 
        num_fullnodes_per_validator: u32, 
        image_tag: &'life1 str, 
        config_overrides: Vec<String>, 
        delete_data: bool
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

#[must_use]
    fn create_validator_and_fullnode_set<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        num_validators: u32, 
        num_fullnodes_per_validator: u32, 
        image_tag: &'life1 str, 
        delete_data: bool
    ) -> Pin<Box<dyn Future<Output = Result<((), ())>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Required methods

#[must_use] fn remove_all_network_effects<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn run<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
instance: &'life1 Instance,
docker_image: &'life2 str,
command: String,
job_name: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Runs the given command in a container against the given Instance

#[must_use] fn validator_instances<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Vec<Instance>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn fullnode_instances<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Vec<Instance>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use] fn upsert_node<'life0, 'async_trait>(
&'life0 self,
instance_config: InstanceConfig,
delete_data: bool
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Inserts an into the ClusterSwarm if it doesn't exist. If it exists, then updates the instance.

#[must_use] fn delete_node<'life0, 'async_trait>(
&'life0 self,
instance_config: InstanceConfig
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Deletes a node from the ClusterSwarm

#[must_use] fn delete_all<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Deletes all validators and fullnodes in this cluster

#[must_use] fn get_grafana_baseurl<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loading content...
Provided methods

#[must_use] fn create_validator_set<'life0, 'life1, 'async_trait>(
&'life0 self,
num_validators: u32,
num_fullnodes_per_validator: u32,
image_tag: &'life1 str,
config_overrides: Vec<String>,
delete_data: bool
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Creates a set of validators with the given image_tag

#[must_use] fn create_fullnode_set<'life0, 'life1, 'async_trait>(
&'life0 self,
num_validators: u32,
num_fullnodes_per_validator: u32,
image_tag: &'life1 str,
config_overrides: Vec<String>,
delete_data: bool
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Creates a set of fullnodes with the given image_tag

#[must_use] fn create_validator_and_fullnode_set<'life0, 'life1, 'async_trait>(
&'life0 self,
num_validators: u32,
num_fullnodes_per_validator: u32,
image_tag: &'life1 str,
delete_data: bool
) -> Pin<Box<dyn Future<Output = Result<((), ())>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Creates a set of validators and fullnodes with the given parameters

Loading content...