Function cluster_test::effects::network_delay::three_region_simulation_effects

pub fn three_region_simulation_effects(
    regions: (Vec<Instance>, Vec<Instance>, Vec<Instance>), 
    delays_bw_regions: (Duration, Duration, Duration)
) -> Vec<NetworkDelay>

three_region_simulation_effects returns the list of NetworkDelays which need to be applied to all the instances in the cluster. regions is a 3-tuple consisting of the list of instances in each region delays_bw_regions is a 3-tuple consisting of the one-way delays between pairs of regions delays_bw_regions.0 is the delay b/w regions 1 & 2, delays_bw_regions.1 is the delay b/w regions 0 & 2, etc