Function cluster_test::experiments::get_experiment

pub fn get_experiment(
    name: &str, 
    args: &[String], 
    cluster: &Cluster
) -> Box<dyn Experiment>

Given an experiment name and its flags, it constructs an instance of that experiment and returns it as a Box<dyn Experiment>