Struct cluster_test::cluster::Cluster

pub struct Cluster { /* fields omitted */ }

Methods

impl Cluster

pub fn from_host_port(peers: Vec<(String, u32)>, mint_file: &str) -> Self

pub fn new_k8s(
validator_instances: Vec<Instance>,
fullnode_instances: Vec<Instance>
) -> Result<Self>

pub fn random_validator_instance(&self) -> Instance

pub fn validator_instances(&self) -> &[Instance]

pub fn fullnode_instances(&self) -> &[Instance]

pub fn all_instances(&self) -> impl Iterator<Item = &Instance>

pub fn into_validator_instances(self) -> Vec<Instance>

pub fn into_fullnode_instances(self) -> Vec<Instance>

pub fn prometheus_ip(&self) -> Option<&String>

pub fn mint_key_pair(&self) -> &KeyPair<Ed25519PrivateKey, Ed25519PublicKey>

pub fn get_validator_instance(&self, name: &str) -> Option<&Instance>

pub fn split_n_validators_random(&self, c: usize) -> (Self, Self)

Splits this cluster into two

Returns tuple of two clusters: First element in tuple contains cluster with c random instances from self Second element in tuple contains cluster with remaining instances from self

pub fn split_n_fullnodes_random(&self, c: usize) -> (Self, Self)

pub fn validator_sub_cluster(&self, ids: Vec<String>) -> Cluster