Module cluster_test::experiments

Structs

CompatibilityTest

CompatiblityTestParams

Context

CpuFlamegraph

CpuFlamegraphParams

LoadTestParams

PacketLossRandomValidators

PacketLossRandomValidatorsParams

PerformanceBenchmark

PerformanceBenchmarkParams

PerformanceBenchmarkThreeRegionSimulation

PerformanceBenchmarkThreeRegionSimulationParams

RebootCluster

RebootClusterParams

RebootRandomValidators

RebootRandomValidatorsParams

Reconfiguration

ReconfigurationParams

RecoveryTime

RecoveryTimeParams

StateSyncPerformance

StateSyncPerformanceParams

TwinValidators

TwinValidatorsParams

ValidatorVersioning

ValidatorVersioningParams

Traits

Experiment

ExperimentParam

Functions

get_experiment
Given an experiment name and its flags, it constructs an instance of that experiment and returns it as a Box<dyn Experiment>