Module cluster_test::experiments

Structs

Context

CpuFlamegraph

CpuFlamegraphParams

PacketLossRandomValidators

PacketLossRandomValidatorsParams

PerformanceBenchmark

PerformanceBenchmarkParams

PerformanceBenchmarkThreeRegionSimulation

PerformanceBenchmarkThreeRegionSimulationParams

RebootRandomValidators

RebootRandomValidatorsParams

RecoveryTime

RecoveryTimeParams

Traits

Experiment

ExperimentParam

Functions

get_experiment
Given an experiment name and its flags, it constructs an instance of that experiment and returns it as a Box<dyn Experiment>