Trait cluster_test::health::HealthCheck

pub trait HealthCheck {
    fn on_event(&mut self, event: &ValidatorEvent, ctx: &mut HealthCheckContext);

    fn name(&self) -> &'static str;

    fn verify(&mut self, _ctx: &mut HealthCheckContext) { ... }

    fn invalidate(&mut self, _validator: &str) { ... }

    fn clear(&mut self) { ... }
}

Required methods

fn on_event(&mut self, event: &ValidatorEvent, ctx: &mut HealthCheckContext)

Verify specific event

fn name(&self) -> &'static str

Loading content...
Provided methods

fn verify(&mut self, _ctx: &mut HealthCheckContext)

Periodic verification (happens even if when no events produced)

fn invalidate(&mut self, _validator: &str)

Optionally marks validator as failed, requiring waiting for at least one event from it to mark it as healthy again

fn clear(&mut self)

Clean is invoked when cluster is wiped This means that checks like commit history check should wipe internal state

Loading content...